+++ src/types.ts
@@ -56,7 +56,7 @@ export type Args = {
 export type Dependency = Record<Metric, number> & {
   dependency: string;
   deprecated?: string;
-  available: string;
+  latest: string;
 };
 
 export type Dependencies = Dependency[];